refactor: dedupe logic to fetch events with txhash
