+++ rust/chains/hyperlane-ethereum/src/contracts/mod.rs
@@ -1,11 +1,8 @@
 pub use {interchain_gas::*, mailbox::*, merkle_tree_hook::*, validator_announce::*};
 
 mod interchain_gas;
-
 mod mailbox;
-
 mod merkle_tree_hook;
-
 mod multicall;
-
+mod utils;
 mod validator_announce;
