+++ rust/chains/hyperlane-ethereum/src/contracts/utils.rs
@@ -1,0 +1,48 @@
+use std::sync::Arc;
+
+use ethers::{
+    abi::RawLog,
+    providers::Middleware,
+    types::{H160 as EthersH160, H256 as EthersH256},
+};
+use ethers_contract::{ContractError, EthEvent, LogMeta as EthersLogMeta};
+use hyperlane_core::{ChainResult, LogMeta, H512};
+use tracing::warn;
+
+pub async fn fetch_raw_logs_and_log_meta<T: EthEvent, M>(
+    tx_hash: H512,
+    provider: Arc<M>,
+    contract_address: EthersH160,
+) -> ChainResult<Vec<(T, LogMeta)>>
+where
+    M: Middleware + 'static,
+{
+    let ethers_tx_hash: EthersH256 = tx_hash.into();
+    let receipt = provider
+        .get_transaction_receipt(ethers_tx_hash)
+        .await
+        .map_err(|err| ContractError::<M>::MiddlewareError(err))?;
+    let Some(receipt) = receipt else {
+        warn!(%tx_hash, "No receipt found for tx hash");
+        return Ok(vec![]);
+    };
+
+    let logs: Vec<(T, LogMeta)> = receipt
+        .logs
+        .into_iter()
+        .filter_map(|log| {
+            // Filter out logs that aren't emitted by this contract
+            if log.address != contract_address {
+                return None;
+            }
+            let raw_log = RawLog {
+                topics: log.topics.clone(),
+                data: log.data.to_vec(),
+            };
+            let log_meta: EthersLogMeta = (&log).into();
+            let event_filter = T::decode_log(&raw_log).ok();
+            event_filter.map(|log| (log, log_meta.into()))
+        })
+        .collect();
+    Ok(logs)
+}
